backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
supabase
python-dotenv
email-validator
scikit-learn

# Web scraping
beautifulsoup4
requests
lxml

# Database & async
sqlalchemy
asyncio
psycopg2-binary

# Scheduling (optional)
APScheduler
